feat: fixed all issues found by sir vaultec,  completely remove the old implementation from the source code as it's no longer necessary and not in use, Just fyi JSON.stringify() is not necessary at all. Just pass in the request body to axios and it'll handle it for you without adding extra headers as well AND LASTLY
suggestion/feature request: We should move ALL this logic here into a Zustand hook so it works globally across the entire frontend. At the moment this essentially creates a scenario where we have potentially multiple requests being triggered across the codebase for the exact same thing.

This would be accomplished through the following:

    Zustand hook for managing state
    Global context to trigger the initial check and only once across the codebase